Rename demoCodeLoader file
