+++ examples/hosts/app-integration/schema-upgrade/src/demoCodeLoader.ts
